requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==64.2.1
djangorestframework==3.15.2
idna==3.9
piexif==1.1.3
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.3
whitenoise==6.7.0
